% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{halfwidthen}
\alias{halfwidthen}
\title{halfwidthen}
\usage{
halfwidthen(x)
}
\arguments{
\item{x}{A string to convert}
}
\description{
Convert Japanese alphanumerics to ASCII alphanumerics
}
\examples{
x = c("\\uff10", "\\uff11")
halfwidthen(x)

}
