% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfwidthr.R
\docType{package}
\name{halfwidthr}
\alias{halfwidthr}
\alias{halfwidthr-package}
\title{halfwidthr}
\description{
This package converts Japanese alphanumerics (a.k.a. "zenkaku") to ASCII alphanumerics (a.k.a. "hankaku").
For example, "１"(0xEFBC91) is converted to "1"(0x31).
}
