% Generated by roxygen2 (4.0.2): do not edit by hand
\name{halfwidthen}
\alias{halfwidthen}
\alias{halfwidthen_one}
\title{halfwidthen}
\usage{
halfwidthen_one(str)

halfwidthen(strs)
}
\arguments{
\item{str}{a string to convert}

\item{strs}{string(s) to convert}
}
\value{
a converted string
}
\description{
{{halfwidthen_one}} takes a single string.
{{halfwidthen}} can take character vector whose length is more than one.
}
\details{
Convert Japanese alphanumerics to ASCII alphanumerics
}
\examples{
x = c("１", "２２Ａ", "３３3")
halfwidthen(x)
halfwidthen_one(x[1])

\dontrun{
# You will see error; halfwidthen_one can take one string as its argument.
halfwidthen_one(x)
}
}

