% Generated by roxygen2 (4.0.2): do not edit by hand
\name{halfwidthr-debug}
\alias{fullwidth_chars}
\alias{halfwidthr-debug}
\alias{hexdump}
\alias{hexinterpret}
\title{halfwidthr-debug}
\usage{
hexdump(s)

hexinterpret(charcode)

fullwidth_chars(start = 15711361L, end = 15711646L)
}
\arguments{
\item{s}{string to hexdump}

\item{charcode}{a charcode between 0x00000000 to 0xFFFFFFFF}

\item{start}{charcode of lower limit of the correspondence table(default: 0xEFBC81, ！)}

\item{end}{charcode of upper limit of the correspondence table(default: 0xEFBC81, ～)}
}
\description{
{{hexdump}} prints given UTF-8 string in hex format.
{{hexinterpret}} prints a UTF-8 chars which the given code point corresponds to.
{{fullwidth_chars}} returns correspondence table between code points and UTF-8 chars.
}
\details{
Functions to dig into UTF-8 byte codes.
}
\examples{
hexdump("あ")
hexdump("月が綺麗ですね")

hexinterpret(0xE38384)

crsp_table <- fullwidth_chars()
crsp_table[1:10,]
}

