% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{halfwidthr}
\alias{halfwidthr}
\alias{halfwidthr-package}
\title{halfwidthr}
\description{
This package converts Japanese alphanumerics (a.k.a. 全角英数) to ASCII alphanumerics (a.k.a. 半角英数).
For example, "１"(0xEFBC91) is converted to "1"(0x31).
}

